Attempt to reorganize and rework gRPC client management.
