+++ service/config/model.go
@@ -18,6 +18,7 @@ type Model struct {
 	Debug bool
 
 	// Mode overrides the endpoint behavior from inference to routing.
+	// This primarily is used to confirm that the mode should be a router.
 	// Can be one of "inference" or "router".
 	// Defaults to "inference".
 	Mode string `json:",omitempty" yaml:",omitempty"`
@@ -35,6 +36,10 @@ type Model struct {
 	Dir string
 
 	// URL is the location of the model.
+	// If Platform is "triton", this is the HTTP prefix of the Triton server, and is deprecated.
+	// It is preferred to use service/endpoint.Config.TritonServers and Triton.ServerID instead.
+	// This will create a new connection per instance URL is used for a Triton server.
+	// Additionally, it is assumed that explicit model control is not enabled for that HTTP URL.
 	URL string
 
 	Batch *BatcherConfigFile `json:",omitempty" yaml:",omitempty"`
@@ -61,6 +66,7 @@ type Model struct {
 	// DataStore is the name of the datastore to use for caching.
 	DataStore string `json:",omitempty" yaml:",omitempty"`
 
+	// Router must be provided if Mode is "router".
 	Router *RouterConfig `json:",omitempty" yaml:",omitempty"`
 
 	// Stream is a github.com/viant/tapper configuration.
@@ -156,7 +162,9 @@ func (m *Model) Validate() error {
 			return fmt.Errorf("triton model %s requires Triton configuration", m.ID)
 		}
 
-		if err := m.Triton.Validate(m.URL != ""); err != nil {
+		m.Triton.Init()
+
+		if err := m.Triton.Validate(m.Mode == "router", m.URL != ""); err != nil {
 			return fmt.Errorf("triton model %s config invalid: %w", m.ID, err)
 		}
 	default:
@@ -176,20 +184,32 @@ func (m *Model) Validate() error {
 	return nil
 }
 
-// TritonConfig represents Triton Inference Server specific configuration
+// TritonConfig represents Triton Inference Server specific configuration.
 type TritonConfig struct {
-	// Model name in Triton
+	// Model name in Triton.
+	// Optional if Model.Mode is "router".
 	ModelName string `json:",omitempty" yaml:",omitempty"`
 
 	// ServerID is the ID of the Triton server.
 	ServerID string `json:",omitempty" yaml:",omitempty"`
 
-	// HTTP timeout in milliseconds
+	// RepositoryExplicit should be true if the Model Repository is in EXPLICIT mode.
+	// In the case of URL-based Triton configuration, the default is to assume POLL mode.
+	// See https://docs.nvidia.com/deeplearning/triton-inference-server/user-guide/docs/user_guide/model_management.html
+	RepositoryExplicit bool `json:",omitempty" yaml:",omitempty"`
+
+	// HTTP total timeout in milliseconds, defaults to 100 milliseconds.
 	Timeout int `json:",omitempty" yaml:",omitempty"`
 }
 
-func (t *TritonConfig) Validate(urlPresent bool) error {
-	if t.ModelName == "" {
+func (t *TritonConfig) Init() {
+	if t.Timeout == 0 {
+		t.Timeout = 100
+	}
+}
+
+func (t *TritonConfig) Validate(isRouter bool, urlPresent bool) error {
+	if !isRouter && t.ModelName == "" {
 		return fmt.Errorf("triton ModelName is required")
 	}
 
@@ -203,7 +223,8 @@ func (t *TritonConfig) Validate(urlPresent bool) error {
 // GetPlatform returns the platform with default to "tensorflow" for backward compatibility
 func (m *Model) GetPlatform() string {
 	if m.Platform == "" {
-		return "tensorflow"
+		m.Platform = "tensorflow"
 	}
+
 	return m.Platform
 }
